feat(181): move connection-string resolution chain/parsers into JD.Efcpt.Build.Core
ConnectionStringResolutionChain (+ context record), AppSettings/AppConfig
parsers, ConfigurationFileTypeValidator, and ConnectionStringResult move
to JD.Efcpt.Build.Core.ConnectionStrings. The context's Log field and
every parser/validator signature flip from the concrete (MSBuild-bound)
BuildLog to IBuildLog, since Core cannot reference Tasks. Chain and
context are public (ResolveSqlProjAndInputs constructs/calls them across
the assembly boundary); the parsers/validator stay internal.

ResolveSqlProjAndInputs (Tasks) repoints its usings at
JD.Efcpt.Build.Core.ConnectionStrings; its own BuildLog-typed private
methods are unaffected since BuildLog implicitly satisfies IBuildLog.

All 3 checked TFMs of Tasks build clean; JD.Efcpt.Build.Tests builds
clean and the 61 targeted ResolutionChain/ConnectionStrings/
ResolveSqlProjAndInputs tests pass unchanged.
