+++ src/JD.Efcpt.Build.Core/ConnectionStrings/AppConfigConnectionStringParser.cs
@@ -1,8 +1,9 @@
 using System.Xml;
 using System.Xml.Linq;
+using JD.Efcpt.Build.Core.Logging;
 using JD.Efcpt.Build.Tasks.Extensions;
 
-namespace JD.Efcpt.Build.Tasks.ConnectionStrings;
+namespace JD.Efcpt.Build.Core.ConnectionStrings;
 
 /// <summary>
 /// Parses connection strings from app.config or web.config files.
@@ -16,7 +17,7 @@ internal sealed class AppConfigConnectionStringParser
     /// <param name="connectionStringName">The name of the connection string to retrieve.</param>
     /// <param name="log">The build log for warnings and errors.</param>
     /// <returns>A result indicating success or failure, along with the connection string if found.</returns>
-    public static ConnectionStringResult Parse(string filePath, string connectionStringName, BuildLog log)
+    public static ConnectionStringResult Parse(string filePath, string connectionStringName, IBuildLog log)
     {
         try
         {
